+++ extensions/openstatus/CHANGELOG.md
@@ -1,3 +1,9 @@
 # OpenStatus Changelog
 
+## [Add "Show Incidents" command] - {PR_MERGE_DATE}
+
+- New "Show Incidents" command
+- "Toggle Details" of a Monitor to see more information
+- Show a `Toast` while _triggering a monitor run_
+
 ## [Initial Version] - 2025-01-15